requirements.txt
python-telegram-bot==21.5
requests==2.32.3
moviepy==1.0.3
sseclient-py==1.7.2
aiohttp==3.9.3
sqlalchemy==2.0.28
aiosqlite==0.20.0
cryptography==42.0.5
python-dotenv==1.0.1
yt-dlp==2023.11.16
openai>=1.50.0
httpx>=0.27.0
ffmpeg